Group Find - /groupfindoff not working
Group Find /groupfind?command=50,100,255 where the ints are r,g,b, lights LEDs until /groupfindoff is sent via a web page.  For some reason Firebeetle does not continue to function or is it just the LCD not updating anymore b/c it will return a value from /read.  This may be an old value though.  I'll figure it out.